apps/800-831.set: new example from Roche FLX
This example is from a camera removed from Roche FLX.

It has a copper data connection instead of fibre.
It wants to talk at 19200 instead of 57600.
It doesn't retain its ADC offsets in non-volatile memory.
It identifies as a model 804 instead of 806.

This file was copied from 800-1134.set, then the baud rate,
model, and serial were updated to the values dumped out of the
camera.  ADC offsets were updated using values that were hand
written on the camera body.