apps/si-dump: if no args, dump interpreted status
Show temperatures and vacuum pressure if no args are
present.

Interpret pressure sensor reading according to S800 user manual.
It needs linear interpolation from a table, so it unfortunately
doesn't map well to Test.cfg for the other apps.

Print in a format suitable for plotting, including the ISO 8601
date.  It may be useful to plot a time series during cool-down
or vacuum pump-out.